refactor(analytics): colocate legacy payload conversion at call boundaries
Remove trackMetaMetricsPayload from the analytics module. Legacy
MetaMetricsEventPayload conversion now lives in MetaMetricsController,
token-detection init, and the metamask-controller UI RPC handler, matching
the smart-transactions init pattern.
