+++ app/scripts/controllers/analytics/index.ts
@@ -11,7 +11,6 @@ export {
   canSubmitAnalytics,
   identify,
   trackEvent,
-  trackMetaMetricsPayload,
   trackPage,
   updateProfileSessionData,
   validateIdentifyPayload,
