+++ app/scripts/controllers/metametrics-controller.test.ts
@@ -48,6 +48,8 @@ import {
   MetaMetricsEventName,
   MetaMetricsUserTrait,
   MetaMetricsUserTraits,
+  type MetaMetricsEventOptions,
+  type MetaMetricsEventPayload,
 } from '../../../shared/constants/metametrics';
 import { CHAIN_IDS } from '../../../shared/constants/network';
 import { KeyringType } from '../../../shared/constants/keyring';
@@ -78,12 +80,13 @@ import {
 import * as analyticsHelpers from './analytics/analytics';
 import {
   configureAnalytics,
+  createEventBuilder,
   getProfileIdentityProperties,
   identify,
-  trackMetaMetricsPayload,
+  trackEvent,
   trackPage,
   updateProfileSessionData,
-} from './analytics/analytics';
+} from './analytics';
 import {
   MetaMetricsController,
   AllowedActions,
@@ -100,6 +103,32 @@ import {
 const TEST_BADGE_FLAG_KEY = 'testTEST338AbtestAttentionBadge';
 const TEST_QUICK_AMOUNTS_FLAG_KEY = 'testTEST4135AbtestQuickAmounts';
 const TEST_LAYOUT_FLAG_KEY = 'testTEST4242AbtestBalanceLayout';
+
+function trackLegacyMetaMetricsPayload(
+  payload: MetaMetricsEventPayload,
+  options?: MetaMetricsEventOptions,
+): void {
+  trackEvent(
+    createEventBuilder(payload.event)
+      .addProperties({
+        ...(payload.properties ?? {}),
+        ...(payload.category === undefined ? {} : { category: payload.category }),
+        ...(payload.revenue === undefined ? {} : { revenue: payload.revenue }),
+        ...(payload.value === undefined ? {} : { value: payload.value }),
+        ...(payload.currency === undefined
+          ? {}
+          : { currency: payload.currency }),
+      })
+      .addSensitiveProperties(payload.sensitiveProperties)
+      .build({
+        environmentType: payload.environmentType,
+        page: payload.page,
+        referrer: payload.referrer,
+        excludeMetaMetricsId: options?.excludeMetaMetricsId,
+        matomoEvent: options?.matomoEvent,
+      }),
+  );
+}
 
 const segmentMock = createSegmentMock(2);
 
@@ -732,7 +761,7 @@ describe('MetaMetricsController', function () {
           analyticsControllerState: { optedIn: false },
         },
         ({ controller }) => {
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Fake Event',
             category: 'Unit Test',
             properties: {
@@ -754,7 +783,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segment, 'track');
           const flushSpy = jest.spyOn(segment, 'flush');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: MetaMetricsEventName.MetricsOptOut,
             category: 'Unit Test',
             properties: {
@@ -790,7 +819,7 @@ describe('MetaMetricsController', function () {
         },
         ({ controller }) => {
           const spy = jest.spyOn(segment, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: MetaMetricsEventName.MetricsOptOut,
             category: 'Unit Test',
           });
@@ -806,7 +835,7 @@ describe('MetaMetricsController', function () {
         },
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Fake Event',
             category: 'Unit Test',
             properties: {
@@ -823,7 +852,7 @@ describe('MetaMetricsController', function () {
     it('should track a legacy event', async function () {
       await withController(({ controller }) => {
         const spy = jest.spyOn(segmentMock, 'track');
-        trackMetaMetricsPayload(
+        trackLegacyMetaMetricsPayload(
           {
             event: 'Fake Event',
             category: 'Unit Test',
@@ -859,7 +888,7 @@ describe('MetaMetricsController', function () {
     it('should track a non legacy event', async function () {
       await withController(({ controller }) => {
         const spy = jest.spyOn(segmentMock, 'track');
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Fake Event',
           category: 'Unit Test',
           properties: {
@@ -897,7 +926,7 @@ describe('MetaMetricsController', function () {
         },
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Fake Event',
             category: 'Unit Test',
             properties: {
@@ -970,7 +999,7 @@ describe('MetaMetricsController', function () {
         },
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Fake Event',
             category: 'Unit Test',
             properties: {
@@ -1003,7 +1032,7 @@ describe('MetaMetricsController', function () {
         expect(() => {
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-expect-error because we are testing the error case
-          trackMetaMetricsPayload({ category: 'test' });
+          trackLegacyMetaMetricsPayload({ category: 'test' });
         }).toThrow(/Must specify event\./u);
       });
     });
@@ -1014,7 +1043,7 @@ describe('MetaMetricsController', function () {
         .mockImplementation(jest.fn());
 
       await withController(async ({ controller }) => {
-        trackMetaMetricsPayload(
+        trackLegacyMetaMetricsPayload(
           {
             event: 'Fake Event',
             category: 'Unit Test',
@@ -1034,7 +1063,7 @@ describe('MetaMetricsController', function () {
     it('tracks sensitiveProperties in a separate event marked for anonymization', async function () {
       await withController(({ controller }) => {
         const spy = jest.spyOn(segmentMock, 'track');
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Fake Event',
           category: 'Unit Test',
           sensitiveProperties: { foo: 'bar' },
@@ -1085,7 +1114,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
 
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Card Button Viewed',
             category: 'Unit Test',
           });
@@ -1132,7 +1161,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
 
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Unified SwapBridge Page Viewed',
             category: 'Unit Test',
           });
@@ -1180,7 +1209,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
 
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Unified SwapBridge Page Viewed',
             category: 'Unit Test',
             properties: {
@@ -1238,7 +1267,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
 
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Card Button Viewed',
             category: 'Unit Test',
             properties: {
@@ -1274,7 +1303,7 @@ describe('MetaMetricsController', function () {
         .mockReturnValue({});
 
       await withController(({ controller }) => {
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Unrelated Event',
           category: 'Unit Test',
         });
@@ -1291,7 +1320,7 @@ describe('MetaMetricsController', function () {
       await withController(({ controller }) => {
         const spy = jest.spyOn(segmentMock, 'track');
 
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Unrelated Event',
           category: 'Unit Test',
           properties: {
@@ -1332,7 +1361,7 @@ describe('MetaMetricsController', function () {
       await withController(({ controller }) => {
         const spy = jest.spyOn(segmentMock, 'track');
 
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Unrelated Event',
           category: 'Unit Test',
           properties: {
@@ -1398,7 +1427,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
 
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Card Button Viewed',
             category: 'Unit Test',
             properties: {
@@ -1468,7 +1497,7 @@ describe('MetaMetricsController', function () {
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
 
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Unified SwapBridge Page Viewed',
             category: 'Unit Test',
           });
@@ -1496,7 +1525,7 @@ describe('MetaMetricsController', function () {
     it('omits profile identity properties when srpSessionData is unavailable', async function () {
       await withController(({ controller }) => {
         const spy = jest.spyOn(segmentMock, 'track');
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Fake Event',
           category: 'Unit Test',
         });
@@ -1520,7 +1549,7 @@ describe('MetaMetricsController', function () {
         } as unknown as MetaMaskState);
 
         const spy = jest.spyOn(segmentMock, 'track');
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Fake Event',
           category: 'Unit Test',
         });
@@ -1571,7 +1600,7 @@ describe('MetaMetricsController', function () {
         } as unknown as MetaMaskState);
 
         const spy = jest.spyOn(segmentMock, 'track');
-        trackMetaMetricsPayload({
+        trackLegacyMetaMetricsPayload({
           event: 'Signature Requested',
           category: 'Unit Test',
           properties: DEFAULT_EVENT_PROPERTIES,
@@ -1609,7 +1638,7 @@ describe('MetaMetricsController', function () {
         },
         ({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload(
+          trackLegacyMetaMetricsPayload(
             {
               event: 'Signature Requested',
               category: 'Unit Test',
@@ -1642,7 +1671,7 @@ describe('MetaMetricsController', function () {
       async (eventType: string) => {
         await withController(({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: eventType,
             category: 'Unit Test',
             properties: DEFAULT_EVENT_PROPERTIES,
@@ -1682,7 +1711,7 @@ describe('MetaMetricsController', function () {
       async (eventType: string) => {
         await withController(({ controller }) => {
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: eventType,
             category: 'Unit Test',
             sensitiveProperties: { foo: 'bar' },
@@ -2696,7 +2725,7 @@ describe('MetaMetricsController', function () {
             TEST_GA_COOKIE_ID,
           );
           const spy = jest.spyOn(segmentMock, 'track');
-          trackMetaMetricsPayload({
+          trackLegacyMetaMetricsPayload({
             event: 'Fake Event',
             category: 'Unit Test',
             properties: {
