+++ app/scripts/messenger-client-init/token-detection-controller-init.ts
@@ -2,8 +2,12 @@ import {
   TokenDetectionController,
   TokenDetectionControllerMessenger,
 } from '@metamask/assets-controllers';
+import type {
+  MetaMetricsEventOptions,
+  MetaMetricsEventPayload,
+} from '../../../shared/constants/metametrics';
 import type { PreferencesControllerState } from '../controllers/preferences-controller';
-import { trackMetaMetricsPayload } from '../controllers/analytics';
+import { createEventBuilder, trackEvent } from '../controllers/analytics';
 import { MessengerClientInitFunction } from './types';
 import { TokenDetectionControllerInitMessenger } from './messengers';
 import { tokenListService } from './token-list-service';
@@ -19,6 +23,36 @@ export const TokenDetectionControllerInit: MessengerClientInitFunction<
       'PreferencesController:getState',
     ) as unknown as PreferencesControllerState;
 
+  const trackMetaMetricsEvent = (
+    payload: MetaMetricsEventPayload,
+    options?: MetaMetricsEventOptions,
+  ) => {
+    trackEvent(
+      createEventBuilder(payload.event)
+        .addProperties({
+          ...(payload.properties ?? {}),
+          ...(payload.category === undefined
+            ? {}
+            : { category: payload.category }),
+          ...(payload.revenue === undefined
+            ? {}
+            : { revenue: payload.revenue }),
+          ...(payload.value === undefined ? {} : { value: payload.value }),
+          ...(payload.currency === undefined
+            ? {}
+            : { currency: payload.currency }),
+        })
+        .addSensitiveProperties(payload.sensitiveProperties)
+        .build({
+          environmentType: payload.environmentType,
+          page: payload.page,
+          referrer: payload.referrer,
+          excludeMetaMetricsId: options?.excludeMetaMetricsId,
+          matomoEvent: options?.matomoEvent,
+        }),
+    );
+  };
+
   const messengerClient = new TokenDetectionController({
     messenger: controllerMessenger,
     disabled: false,
@@ -27,7 +61,7 @@ export const TokenDetectionControllerInit: MessengerClientInitFunction<
         'AssetsContractController:getBalancesInSingleCall',
         ...args,
       ),
-    trackMetaMetricsEvent: trackMetaMetricsPayload,
+    trackMetaMetricsEvent,
     useTokenDetection: () => Boolean(getRetypedPrefState().useTokenDetection),
     // Don't reach external services (token list API + balance multicall) until
     // onboarding is complete — otherwise detection triggered by the vault
